Refactored firebaseLoading. Removed console warns
