+++ src/index.js
@@ -1,4 +1,5 @@
 import logMessage from './js/logger';
 import './css/style.css';
+//import * as firebase from './js/firebaseLoad';
 // Log message to console
 logMessage('Welcome to Expack!');
