+++ src/js/edit.js
@@ -1,26 +1,13 @@
 //These imports are needed for the program to run properly
 import logMessage from './logger';
 import '../css/quantity.css';
-var firebase = require('firebase');
+
+import {getDatabaseReference} from './firebaseLoad';
 // Log message to console
-logMessage('Welcome to Data Check!');
-
-//All of the firebase information needed to interact with the database
-var firebaseConfig = {
-    apiKey: "AIzaSyCK_wBNL7Fhpj7ZC0cDlZ3EhnTvbbYiE24",
-    authDomain: "yerc-rivernet.firebaseapp.com",
-    databaseURL: "https://yerc-rivernet.firebaseio.com",
-    projectId: "yerc-rivernet",
-    storageBucket: "yerc-rivernet.appspot.com",
-    messagingSenderId: "634881168606",
-    appId: "1:634881168606:web:14164f57364baa842d9e4f",
-    measurementId: "G-FH7MS1FE95"
-};
-
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
 
 import { validateInput, validateAllQuality, validateDate} from "./util";
+
+logMessage('Welcome to Data Check!');
 
 //The purpose of this function is to see what jar the user wants to look at
 // and then pulls up all that info in text boxes so the data can be evaluated and edited
@@ -38,7 +25,7 @@ function viewData(){
     }
     
     //Used to access the data in the database at a specefic moment
-    return firebase.database().ref().once('value').then(function(snapshot) {
+    return getDatabaseReference().ref().once('value').then(function(snapshot) {
         
         //These two lines are designed to convert all the keys of the database into an array
         var allDateSnapshot = snapshot.val();
